refactor: move AnalysisConfig/AgentAnalysisConfig to config module
Decouple config from analyzer by moving pure data types
(AnalysisConfig, AgentAnalysisConfig) into src/config/analysis.rs.
Analyzer re-exports them to preserve the public API.

Also adds cargo-husky pre-commit hook for cargo fmt and skips
config::migrate in Miri CI (insta snapshots spawn processes).
